removing unwanted files from commit
