--- ingestion/functions/parsing/peru/input_event.json
@@ -1,8 +1,0 @@
-{
-    "env": "local",
-    "s3Bucket": "epid-sources-raw",
-    "s3Key": "test/peru/Peru_positivos_covid_jan14_head5k.csv",
-    "sourceUrl": "https://cloud.minsa.gob.pe/s/Y8w3wHsEdYQSZRp/download",
-    "sourceId": "60006af8d22cdc005c3f637c"
-}
-
