--- ingestion/functions/parsing/republic_of_korea/input_event.json
@@ -1,8 +1,0 @@
-{
-    "env": "local",
-    "s3Bucket": "epid-sources-raw",
-    "s3Key": "test/korea_jan5.csv",
-    "sourceUrl": "https://raw.githubusercontent.com/ThisIsIsaac/Data-Science-for-COVID-19/master/Covid19_Dataset/patients.csv",
-    "sourceId": "5ffc6a9e7a354a0038b37dfc"
-}
-
