+++ ingestion/functions/parsing/scotland/input_event.json
@@ -1,0 +1,7 @@
+{
+    "env": "local",
+    "s3Bucket": "epid-sources-raw",
+    "s3Key": "test/scotland/scotland_covid_trend_agesex_20210115.csv",
+    "sourceUrl": "https://www.opendata.nhs.scot/dataset/b318bddf-a4dc-4262-971f-0ba329e09b87/resource/9393bd66-5012-4f01-9bc5-e7a10accacf4/download/trend_agesex_$FULLYEAR$FULLMONTH$FULLDAY.csv",
+    "sourceId": "6002b302b13c8100365ac7d9"
+}
