+++ ingestion/functions/parsing/scotland/sample_data.csv
@@ -1,0 +1,21 @@
+,Date,Country,Sex,SexQF,AgeGroup,AgeGroupQF,DailyPositive,CumulativePositive,CrudeRatePositive,DailyDeaths,CumulativeDeaths,CrudeRateDeaths,CumulativeNegative,CrudeRateNegative
+40,20200229,S92000003,Total,d,15 to 19,,0,0,0.0,0,0,0.0,19,6.738592272607981
+41,20200229,S92000003,Total,d,20 to 24,,0,0,0.0,0,0,0.0,70,20.146435807699397
+42,20200229,S92000003,Total,d,25 to 44,,0,0,0.0,0,0,0.0,276,19.3864878988981
+43,20200229,S92000003,Total,d,45 to 64,,0,1,0.0666504928137439,0,0,0.0,301,20.0617983369369
+44,20200229,S92000003,Total,d,65 to 74,,0,0,0.0,0,0,0.0,74,12.796126577900699
+45,20200229,S92000003,Total,d,75 to 84,,0,0,0.0,0,0,0.0,14,4.13274372856139
+46,20200229,S92000003,Total,d,85plus,,0,0,0.0,0,0,0.0,3,2.3605876289471
+47,20200229,S92000003,Total,d,Total,d,0,1,0.0183039554847803,0,0,0.0,863,15.7963135833654
+48,20200301,S92000003,Female,,15 to 19,,0,0,0.0,0,0,0.0,11,7.983394539358129
+49,20200301,S92000003,Female,,20 to 24,,0,0,0.0,0,0,0.0,48,28.006301417819
+50,20200301,S92000003,Female,,25 to 44,,0,0,0.0,0,0,0.0,134,18.598763598610898
+51,20200301,S92000003,Female,,45 to 64,,0,0,0.0,0,0,0.0,153,19.767824957848
+52,20200301,S92000003,Female,,65 to 74,,1,1,0.331673858461498,0,0,0.0,37,12.2719327630754
+53,20200301,S92000003,Female,,75 to 84,,0,0,0.0,0,0,0.0,8,4.18795544015412
+54,20200301,S92000003,Female,,85plus,,0,0,0.0,0,0,0.0,4,4.86653527021437
+55,20200301,S92000003,Female,,Total,d,1,1,0.0357104978507637,0,0,0.0,447,15.962592539291402
+56,20200301,S92000003,Male,,15 to 19,,0,0,0.0,0,0,0.0,11,7.62977554587576
+57,20200301,S92000003,Male,,20 to 24,,0,0,0.0,0,0,0.0,30,17.0390648961185
+58,20200301,S92000003,Male,,25 to 44,,0,0,0.0,0,0,0.0,158,22.468906162453
+59,20200301,S92000003,Male,,45 to 64,,1,2,0.27533835642274895,0,0,0.0,162,22.302406870242702
