+++ ingestion/functions/template.yaml
@@ -440,6 +440,19 @@ Resources:
         - AWSLambdaReadOnlyAccess
       Layers:
         - !Ref ParsingLibLayer
+  ScotlandParsingFunction:
+    Type: AWS::Serverless::Function # More info about Function Resource: https://github.com/awslabs/serverless-application-model/blob/master/versions/2016-10-31.md#awsserverlessfunction
+    Properties:
+      CodeUri: parsing/scotland/
+      Handler: scotland.lambda_handler
+      Description: Parse case data for Scotland
+      # Function's execution role
+      Policies:
+        - AWSLambdaBasicExecutionRole
+        - AWSLambdaReadOnlyAccess
+      Layers:
+        - !Ref ParsingLibLayer
+
 
 
 # Declare values that can be imported to other CloudFormation stacks, or should
